+++ django/mysite/mysite/settings/defaults.py
@@ -32,7 +32,6 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     'unicorn.apps.UnicornConfig',
-    'unicorn_api.apps.UnicornApiConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
