+++ django/mysite/mysite/urls.py
@@ -21,8 +21,6 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
-    url(r'^', include('unicorn.urls')),
-    url(r'^unicorn_api/', include('unicorn_api.urls')),
-    url(r'^login/', views.login, name='login'),
+    url(r'^api/', include('unicorn.urls')),
     url(r'^admin/', admin.site.urls),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
