--- django/mysite/unicorn/templates/unicorn/article_detail.html
@@ -1,42 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-<h1>{{ article.headline }}</h1>
-
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-2">
-			<h3>Authors</h3>
-		</div>
-		<div class="col-md-10">
-		{% for author in article.authors.all %}
-			<h4>{{author.last_name}} , {{author.first_name}}</h4>
-		{% endfor %}
-		</div>
-	</div>
-	<div class="row">
-		<div class="col-md-2">
-			<h3>Tags</h3>
-		</div>
-		<div class="col-md-10">
-			{% for tag in article.tags.all %}
-				<h4>{{tag.text}}</h4>
-			{% endfor %}
-		</div>
-	</div>
-	<div class="row">
-		<div class="col-md-2">
-			<h3>Copy</h3>
-		</div>
-		<div class="col-md-10">
-			<p>{{article.copy}}</p>
-		</div>
-	</div>
-</div>
