--- django/mysite/unicorn/templates/unicorn/article_list.html
@@ -1,43 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-8">
-			<h1>Article List</h1>
-			{% if article_list %}
-			    <ul>
-				    {% for article in article_list %}
-					    <li>
-					    	<h3><a href="/articles/{{ article.slug }}/">{{ article.headline }}</a></h3>
-					    	<ul>
-					    		<li>Tags</li>
-					        {% for tag in article.tags.all %}
-								<ul>
-									<li>{{tag.text}}</li>
-								</ul>
-					    	{% endfor %}
-							</ul>
-							<ul>
-					    		<li>Authors</li>
-					        {% for author in article.authors.all %}
-								<ul>
-									<li>{{author.last_name}}, {{author.first_name}}</li>
-								</ul>
-					    	{% endfor %}
-							</ul>
-					    </li>
-				    {% endfor %}
-			    </ul>
-			{% else %}
-			    <p>No Articles are available.</p>
-			{% endif %}
-		</div>
-	</div>
-</div>
