--- django/mysite/unicorn/templates/unicorn/author_detail.html
@@ -1,20 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-<h1>{{author.first_name}} {{author.last_name}}</h1>
-
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-2">
-			<h3>{{author.bio}}</h3>
-			<h6>{{author.school}}</h6>
-		</div>
-	</div>
-</div>
