--- django/mysite/unicorn/templates/unicorn/author_list.html
@@ -1,23 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-<div class="container-fluid">
-	<div class="row">
-		<div class="col-md-8">
-			<h2>Author List </h2>
-			{% if author_list %}
-				{% for author in author_list%}
-					<h4><a href="/authors/{{author.pk}}">{{author.last_name}}, {{author.first_name}}</a></h4>
-				{% endfor %}
-			{% else %}
-				<p>No Authors are available</p>
-			{% endif%}
-		</div>
-	</div>
-</div>
