--- django/mysite/unicorn/templates/unicorn/home.html
@@ -1,11 +1,0 @@
-{# Load the tag library #}
-{% load bootstrap3 %}
-
-{# Load CSS and JavaScript #}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
-
-{# Display django.contrib.messages as Bootstrap alerts #}
-{% bootstrap_messages %}
-
-<h1><em>thecavalierdani</em>Unicorn</h1>
