--- django/mysite/unicorn_api/apps.py
@@ -1,7 +1,0 @@
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class UnicornApiConfig(AppConfig):
-    name = 'unicorn_api'
